Ensure the Player's listener is zero'd out as part of Player::Destroy(). Also cleaned up a couple small warnings.

+++ src/core/audio/Crossfader.cpp
@@ -165,7 +165,7 @@ void Crossfader::ProcessMessage(IMessage &message) {
                     percent = (1.0f - percent);
                 }
 
-                float outputVolume = globalVolume * percent;
+                double outputVolume = globalVolume * percent;
 
 #if 0
                 std::string dir = (fade->direction == FadeIn) ? "in" : "out";
